+++ src/main/java/com/example/digitalmenu/dtos/FoodResponseDTO.java
@@ -2,7 +2,9 @@ package com.example.digitalmenu.dtos;
 
 import com.example.digitalmenu.models.Food;
 
-public record FoodResponseDTO(Long id, String title, String image, Double price) {
+import java.math.BigDecimal;
+
+public record FoodResponseDTO(Long id, String title, String image, BigDecimal price) {
     public FoodResponseDTO(Food food) {
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
     }
